requirements.txt
fastapi==0.109.0 
uvicorn[standard]==0.27.0 
python-dotenv==1.0.0 
 
langchain==0.1.0 
langchain-huggingface==1.1.0
langchain-community==0.0.10 
huggingface-hub==0.20.2 
transformers==4.36.2 
accelerate>=0.26.0
unstructured==0.10.30
"unstructured[md]"
sentence-transformers==2.2.2
# Vector DB 
chromadb==0.4.22 
 
# Messaging 
paho-mqtt==1.6.1 
 
# Testing 
pytest==7.4.4 
httpx==0.26.0 
